+++ apps/src/componentLibrary/dropdown/_CustomDropdown.tsx
@@ -29,6 +29,8 @@ export interface CustomDropdownProps {
   size: ComponentSizeXSToL;
   /** CustomDropdown disabled state */
   disabled?: boolean;
+  /** CustomDropdown readOnly state */
+  readOnly?: boolean;
   /** CustomDropdown label
    * The user-facing label of the dropdown */
   labelText: string;
@@ -55,6 +57,7 @@ const CustomDropdown: React.FunctionComponent<CustomDropdownProps> = ({
   isSomeValueSelected = false,
   icon,
   disabled = false,
+  readOnly = false,
   color = dropdownColors.black,
   size = 'm',
 }) => {
@@ -118,12 +121,15 @@ const CustomDropdown: React.FunctionComponent<CustomDropdownProps> = ({
       <button
         id={`${name}-dropdown-button`}
         type="button"
-        className={moduleStyles.dropdownButton}
+        className={classNames(
+          moduleStyles.dropdownButton,
+          readOnly && moduleStyles.readOnlyDropdownButton
+        )}
         data-toggle="dropdown"
         aria-haspopup={true}
         aria-label={`${name} filter dropdown`}
         onClick={toggleDropdown}
-        disabled={disabled}
+        disabled={disabled || readOnly}
       >
         {isSomeValueSelected && (
           <FontAwesomeV6Icon iconName="check-circle" iconStyle="solid" />