+++ apps/src/componentLibrary/dropdown/checkboxDropdown/CheckboxDropdown.tsx
@@ -31,6 +31,8 @@ export interface CheckboxDropdownProps {
   size: ComponentSizeXSToL;
   /** CheckboxDropdown disabled state */
   disabled?: boolean;
+  /** CheckboxDropdown readOnly state */
+  readOnly?: boolean;
   /** CheckboxDropdown label
    * The user-facing label of the dropdown */
   labelText: string;
@@ -71,6 +73,7 @@ const CheckboxDropdown: React.FunctionComponent<CheckboxDropdownProps> = ({
   onSelectAll,
   onClearAll,
   disabled = false,
+  readOnly = false,
   color = dropdownColors.black,
   size = 'm',
 }) => {
@@ -81,6 +84,7 @@ const CheckboxDropdown: React.FunctionComponent<CheckboxDropdownProps> = ({
       labelType={labelType}
       color={color}
       disabled={disabled}
+      readOnly={readOnly}
       size={size}
       isSomeValueSelected={checkedOptions.length > 0}
     >