+++ apps/src/componentLibrary/dropdown/checkboxDropdown/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.4.0](https://github.com/code-dot-org/code-dot-org/pull/58209)
+* added `readOnly` prop to allow for setting the dropdown to read-only
+* minor styles updates for consistency
+
 ## [0.3.0](https://github.com/code-dot-org/code-dot-org/pull/57827)
 * added `thick` and `thin` label style types (`labelType` prop)
 * added black and white background color fill